requirements.txt
fastapi>=0.115.0
httpx>=0.28.0
Pillow>=10.0.0
pytest>=8.0.0
python-multipart>=0.0.20
torch
torchvision
uvicorn[standard]>=0.34.0
